Migrate build tooling from Vue CLI to Vite
Replace Vue CLI (webpack) with Vite for faster dev/builds and a leaner
dependency tree, and optimize the homepage portrait image.

Build tooling:
- Add vite.config.js (@vitejs/plugin-vue, @ alias, base '/', modern Sass API)
- Move public/index.html to project root as the Vite entry (module script,
  drop webpack template tokens)
- Swap vue-cli/babel/core-js deps for vite + @vitejs/plugin-vue; add dev/
  preview scripts; pin node-fetch (used by the scraper) explicitly
- Remove vue.config.js, babel.config.js, .browserslistrc
- Rename scripts/gh-scraper.js -> .cjs (package is now ESM) and update the
  scheduler workflow

Asset paths:
- Normalize all public asset references (img/, data/) to leading-slash
  absolute paths so they resolve on every route under Vite, fixing the
  previously broken /public/img/ background images in AboutView

Portrait image:
- Replace the 16 MB 3370x3370 cover_image.png (shown at 30vw) with a
  135 KB WebP plus a resized PNG fallback via <picture>, with
  fetchpriority="high"
